cva: Initial commit
Some projects use virtual machines in order to increase the security
between containers. They need a way to communicate with the VM to be
able to interact with containers running inside it.

This project aims to provide an application running inside a VM and
responsible for managing containers lifecycle. It defines a protocol
so that anyone can send some commands and get control over the
containers running inside the VM.

It relies on the libcontainer provided by runc, making the container
management easier, and reducing the amount of code needed.

Notice that "cva" stands for "container-vm-agent".

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>