network: Support more IP and netmast formats
In order to get Clear Containers 2.1 working, we have to accept
slightly different formats from what we were previously expecting
regarding the netmask and the destination IP of a route.

About the routes, we now accept the destination IP without the
netmask attached, meaning the agent has to find it and concatenate
it to the IP provided.

About the interfaces, if the netmask is in IP format 255.255.0.0
for instance, the agent will ignore it and will rely on the IP
address to determine the netmask in decimal format.

Fixes #40

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>