agent: DestroyPod has to terminate the agent
This patch simplifies what we tried to do earlier to terminate
properly each goroutine listening on both CTL and TTY channels.

It basically exits the CTL channel go routine in case we received
a DESTROYPOD command (after it sent ACK or ERROR).

The main now terminates if one of the two listening routines returns
for any reason. That way, no need to sync between CTL and TTY.

Fixes #42

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>