channels: Stat /sys/class/virtio-ports/*/name before reading it
We don't want to end up getting an error because we tried to read a
non existing file. Indeed, we want to scan every entry inside the
path /sys/class/virtio-ports/, without throwing an error because
on of those entries does not have a file named "name". In that
case, we just want to skip it.

Fixes #58

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>