agent: Kill only init process if specified this way
In case we were running CRI-O testing inside Semaphore, because the
system is kind of slow, the redis container process takes some time
to really start. Because we were issuing a kill signal right after
the start returned (does not mean the container's process finished
to start), and at the same time we were trying to kill all processes
related to the init process, this was hanging. Indeed, we should not
try to kill all processes until the init one is properly running.

This patch relies on AllProcesses flag provided in the "killcontainer"
payload, instead of hardcoding this value to "true". Indeed, the CRI-O
tests explicitly ask for a kill without the --all flag enabled. This
simple change allows our agent to get CRI-O testing passing properly
on Semaphore CI.

Fixes #61

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>