sync: Synchronize container map in pod struct
Maps in golang are not safe for concurrent use.
Although the serial port that is used by the agent translates
to some degree of serialization, we should protect the containers
map from access by several go-routines.
There is a chance that removecontainer may be called while
a container is cleaning up its process pipes.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>