agent: Refactor closeProcessStreams/Pipes to be called on container
The functions closeProcessStreams() and closeProcessPipes were being
called on pod struct pointer. Refactor these to be called on
container instead. Semantically it makes sense to have these
methods called on a container.
This also cleans up the code as we avoid accesses to container map
in the pod to fetch a container in order to access the container's
processes map.

The next step would be to protect the processes maps in container
struct for concurrent acceses as these are very much likely
to be accessed concurrently when several processes in a container
are exiting.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>