agent: Update libcontainer and remove the reaper fix
The latest version of libcontainer introduces a fix to reap the
defunct child process so that we don't have to do that from our
agent anymore.

This is pretty logical since the lib is starting the process on
our behalf, therefore they know they are leaving a zombie process
behind us.

Fixes #88

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>