kill: Prevent deadlock while waiting for kill command
Even if the source of our kill deadlock bug has been solved through
libcontainer directly, we might want to avoid any potential lock
when issuing a signal to a container. That's why this patch introduces
a timeout after which we consider the kill command as failing.
Moreover, it will show us easily if we're getting caught in such a case
in the future, by printing important logs.

Fixes #130

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>