agent: Discard signals if cannot find sequence ID
When a container with terminal enabled is started, it is very likely
that it's going to receive a SIGWINCH signal right after it started.
But there is an issue here, if the workload exited very quickly, when
the signal arrives to the agent, it cannot find the process/seqID.

The race condition happens because at the same time, the agent sends
the exit code to the shim, and the shim sends the signal to the agent.

The way to avoid such issues is to proceed the same way it was done
with the sdtin coming from IO channel, that is dropping the signal
if we have no seqID match. That's why this patch discards signals,
either because there is no corresponding seqID, or because the
container holding those seqID has been stopped. In both cases, we
cannot prevent the agent to receive signals, and the best we can do
is to ignore them.

Fixes #142

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>