ci: Standardize agent timeouts to 15 seconds
In order to avoid hitting timeouts when our CI runs on top of virtual
machines, we should increase all timeouts to an acceptable default
value.

In this particular case, we need to bump 'newContainer' and 'exec'
timeouts from 5 to 15 seconds. The 'kill' command timeout was already
set up to 15 seconds, this will become the standard timeout for the
whole agent.

Fixes #148

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>