agent-cli: Allow reconnection to the agent
Because the agent sends a READY message the first time a connection
is established, there is no way to reconnect without restarting the
agent. In order to workaround this, a new flag is added to the agent
CLI so that we can make a difference between those two cases.

Fixes #150

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>