network: Update route list according to what's on the system
When using the destroyPod() function, the removal of the network is
performed but it generates an error. The error says that we cannot
remove the route. This is caused by the fact that we should not try
to remove this route since we have not added it. Indeed, when we do
the network setup, we skip the existing routes.

This patch removes routes already on the system from the route list
held in the pod structure. That way, when destroyPod() is called,
the list does not contain the routes that we have not added.

Fixes #157

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>