destroy: Change DESTROYPOD behavior not to exit
This patch is simple but it modifies the complete behavior of the
DESTROYPOD command since it won't cause the agent process to exit
and it won't cause the VM to exit subsequently.

Indeed, we expect the runtime to be responsible for destroying the
VM, and not the agent itself. Moreover, this behavior could allow
the support for a restart of the pod without restarting the VM.
The future will tell us if that's something we need to implement.

Fixes #159

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>