stdin: Close process stdin when receiving EOF
In case the shim stdin is closed, it is going to receive an empty
message when reading from it. This behavior has to be reproduced
on the agent side so that it can determine when to close stdin.

The proxy will forward an empty message and this patch will make
the agent aware that receiving an empty message for a valid seq
ID should be processed as an EOF. In that case, stdin will be
closed and the process will end up properly (if it was waiting
on the input to finish).

Fixes #161

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>