stdin: Do not close stdin if this is a terminal
This case should not happen since this is not a valid case, but it
is better to ensure we check this directly from the agent. In case
a terminal has been set, closing the stdin would close the terminal,
which would lead the runnning process to not be able to send any
output, leading to errors on the agent.

The case where we need to close stdin has to happen only for non
terminal case. That way stdout and stderr will still be open and
outputs will continue to come from the process.

Fixes #161

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>