agent: Make the agent the subreaper of all processes
Our agent had a flaw, it potentially lives with some zombie processes
according to what has been done by the container user. The explanation
is that we only wait for container and exec processes but we don't
handle the children and grandchildren processes coming from those
initial container and exec processes. The example of such a case can
be shown demonstrated in case the container process spawn a child
process, and this one will also spawn another child process. In case
the child terminates before its own children, those ones expect to be
reaped by the init process of the PID namespace, i.e the container
process. But this container process is not written to reap descendant
processes, which will lead to a zombie process left behind.

This commit solves those cases by setting the agent process, who will
be the father of all container processes, as a subreaper. This way, it
will receive all the SIGCHLD signals from any child left behind, and
will reap them. Notice this patch makes sure we don't check the error
from the libcontainer call to process.Wait() since it won't be able to
reap the container or exec processes as expected. This libcontainer
Wait() function is only used here to properly cleanup the pipes and
internal structures created by libcontainer.

Fixes #177

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>