dev: Mount /dev from tmpfs before mount namespace is created
We use tmpfs to mount a container's /dev so that the device gets
access to a limited set of devices, but this means that the container
cannot access devices that have been created within the container.

We need to mount newly created devices to a container's /dev so that
the container can get access to these devices.

/dev was being mounted by libcontainer in container's mount namespace
followed by a chroot into the container's rootfs.
As this is a newly created mount point, it is not possible to access
a container's /dev from the agent mount namespace.

Even if we do a setns using something like docker's nsenter cgo package,
this would mean we would not be able to access the outside devices
because of the chroot created before the libcontainer call for starting
a container returns.

To overcome this, set up /dev for a container within the agent so
that the mount point is shared with the agent's mount namespace.
Note, this still uses tmpfs so that we get an initial limited set of
devices within the container.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>