debug: Send errors on the control channel
We were always sending the response that was received on executing agent
callback function for a command, on the control channel.
In case of an error, this response is empty and hence the error is not
propogated to the proxy at all, resulting in an empty message sent to
the proxy in case of error.
Instead check for error case and send the error string on the control
channel. With this, we can finally see the errors occurring in the agent
in the proxy logs.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>