devices: Bind mount devices only when /dev is expected to be bindmounted
As per runc behaviour new devices are expected to show up when
-v /dev:/dev is passed on the docker command line. In this case, docker
passes /dev as a bind mount rather than a tmpfs based mount.

This really means that the host /dev should be replicated inside the container.
Since that is not possible in case of VM based container's, we do our
best by mounting the newly created devices while interacting with the
container.

We pass a bool "BindMountDev" to newContainer payload to flag this case. New
devices are bind-mounted only when this flag is true.
This flag needs to be passed by the runtime.

We also get rid of the VFIO bind mount, which was an interim solution.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>