scsi: Add support to pass SCSI address for SCSI disks
Add support for SCSI disks to be passed for container rootfs.
If SCSI address is provided, use that to get the SCSI disk name.
This eliminates the need to predict the drive name of the block
device on the host side. We do need to incur the cost of scanning SCSI bus.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>